Update mean_ap.py (#41)

+++ visionmetrics/detection/mean_ap.py
@@ -61,6 +61,6 @@ class MeanAveragePrecision(detection.mean_ap.MeanAveragePrecision):
             boxes = torch.tensor(boxes) if not isinstance(boxes, torch.Tensor) else boxes
 
         if scores:
-            return {'boxes': boxes[:, -4:], 'labels': boxes[:, 0].to(torch.int), 'scores': boxes[:, 1]}
+            return {'boxes': boxes[:, -4:], 'labels': boxes[:, 0].to(torch.int), 'scores': boxes[:, 1].to(torch.float)}
         else:
             return {'boxes': boxes[:, -4:], 'labels': boxes[:, 0].to(torch.int)}
